[chore] overall edit css and structure
