+++ src/pages/StudyCategoryPage.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { useLocation, useNavigate } from 'react-router-dom';
 import axios from 'axios';
+import Button from '../components/Button';
 import SearchBar from '../components/SearchBar';
 
 const RecruitBox = styled.div`
@@ -99,7 +100,7 @@ const Title = styled.div`
 `;
 
 const ChallengeNumber = styled.div`
-  margin-left: 210px;
+  margin-left: 230px;
   margin-top: 10px;
   font-size: 1.1rem;
   position: fixed;
@@ -107,7 +108,7 @@ const ChallengeNumber = styled.div`
 
 const Line = styled.div`
   margin-left: 10px;
-  width: 280px;
+  width: 295px;
   border-bottom: 3px solid #000;
   position: fixed;
 `
@@ -119,7 +120,7 @@ const Body = styled.div`
   overflow-y: auto;
 `;
 
-const ImminentPage = () => {
+const CategoryPage = () => {
     const navigate = useNavigate();
     const [recruit, setRecruit] = useState([]);
     const [total, setTotal] = useState("0");
@@ -134,7 +135,7 @@ const ImminentPage = () => {
     };
 
     const getRecruit = () => {
-        axios.get(`http://43.200.19.7:8080/api/v1/main/imminent/all`,  {
+        axios.get(`http://43.200.19.7:8080/api/v1/main/category?category=자유스터디`,  {
             headers: {
                 'Content-Type': 'application/json',
                 'Authorization': ` Bearer ${ACCESS_TOKEN}`
@@ -154,7 +155,7 @@ const ImminentPage = () => {
     return (
         <RecruitBox>
             <TitleBox>
-                <Title>마감 임박 챌린지</Title>
+                <Title>자유스터디 챌린지</Title>
                 <ChallengeNumber>
                     (총 {total}개)
                 </ChallengeNumber>
@@ -171,20 +172,18 @@ const ImminentPage = () => {
                     <RecruitInfo>
                         <RecruitTitle>{challenge.challenge_title}</RecruitTitle>
                         <RecruitDetails>
-                            <span>기간</span>&nbsp;&nbsp;
-                            <span style={{color:"#42AF53"}}>{challenge.startDate.substring(0, 4)}.{challenge.startDate.substring(4, 6)}.{challenge.startDate.substring(6, 8)}</span>
+                            <span>기간</span>
+                            <span style={{fontWeight:'bold'}}>{challenge.startDate.substring(0, 4)}.{challenge.startDate.substring(4, 6)}.{challenge.startDate.substring(6, 8)}
                             &nbsp;~&nbsp;
-                            <span>{challenge.endDate.substring(0, 4)}.{challenge.endDate.substring(4, 6)}.{challenge.endDate.substring(6, 8)}</span>
-                        </RecruitDetails>
+                            {challenge.endDate.substring(0, 4)}.{challenge.endDate.substring(4, 6)}.{challenge.endDate.substring(6, 8)}</span></RecruitDetails>
                         <RecruitExplain>{challenge.detail}</RecruitExplain>
                     </RecruitInfo>
                     </div>
                 ))}
                 </RecruitList>
             </Body>
-        
         </RecruitBox>
     );
 };
 
-export default ImminentPage;
+export default CategoryPage;
