+++ src/pages/TestCategoryPage.jsx
@@ -3,106 +3,121 @@ import styled from 'styled-components';
 import { useLocation, useNavigate } from 'react-router-dom';
 import axios from 'axios';
 import Button from '../components/Button';
+import SearchBar from '../components/SearchBar';
 
 const RecruitBox = styled.div`
-    margin-left: 4vw;
-    margin-right: 3vw;
-    padding-top: 1.2vw;
-    font-family:'Pretendard';
-    margin-bottom: 200px;
+  margin-left: 47.5px;
+  margin-right: 40px;
+  padding-top: 10px;
+  font-family:'Pretendard';
+  margin-bottom: 180px;
 `;
 
 const RecruitList = styled.div`
   display: grid;
-  grid-template-columns: 25% 25% 25% 25%;
-  justify-content: space-between;
+  grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
+  gap: 60px;
   width: 100%;
-  margin-bottom: 50px;
 `;
 
 const RecruitImageContainer = styled.div`
   border: 2px solid #ffd700;
   border-radius: 30px;
   overflow: hidden;
-  width:253px;
+  width: 253px;
   height:347px;
   cursor: pointer;
 `;
 
 const RecruitImage = styled.img`
-    width:100%;
-    height: 100%;
-    border-bottom: 1px solid #ccc;
-    object-fit: cover;
+  width:100%;
+  height: 100%;
+  border-bottom: 1px solid #ccc;
+  object-fit: cover;
 `;
 
 const RecruitInfo = styled.div`
-    width:17rem;
-    margin-bottom: 100px;
 `;
 
 const RecruitTitle = styled.h3`
-    font-weight: bold;
-    font-size:1.25rem;
+  font-weight: 600;
+  font-size: 1.2rem;
+  color: black;
+  margin-bottom: 5px;
+  width: 245px;
+  text-align: left;
+  word-wrap: break-word;
+  line-height: 1.2;
+  display: -webkit-box;
+  -webkit-line-clamp: 1;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
 `;
 
 const RecruitDetails = styled.div`
-    margin-top: 0.5rem;
-    border-top: 1px solid #ccc;
-    padding-top: 0.5rem;
-    font-size:1.1rem;
-    display: flex;
-    justify-content: space-between;
-`;
-const RecruitExplain = styled.div`
-    margin-top: 0.5rem;
-    border-top: 1px solid #ccc;
-    padding-top: 0.5rem;
-    font-size:1.1rem;   
-    text-overflow: ellipsis;
-    overflow: hidden;
-    word-break: break-word;
-    width: 17rem;
-    display: -webkit-box;
-    -webkit-line-clamp: 2; // 원하는 라인수
-    -webkit-box-orient: vertical;
-`
-const BtnContainer = styled.div`
-  position: fixed;
-  top: 75%;
-  right: -20px;
+  width: 245px;
+  font-weight: 600;
+  font-size: 1.1rem;
+  margin-top: 0.5rem;
+  border-top: 1px solid #ccc;
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  display: flex;
+  justify-content: space-between;
 `;
 
+const RecruitExplain = styled.div`
+  width: 245px;
+  border-top: 1px solid #ccc;
+  padding-top: 0.5rem;
+  font-size: 1rem;   
+  text-overflow: ellipsis;
+  overflow: hidden;
+  word-break: break-word;
+  display: -webkit-box;
+  -webkit-line-clamp: 2; // 원하는 라인수
+  -webkit-box-orient: vertical;
+`
+
 const TitleBox = styled.div`
-    margin-right: 2vw;
-    margin-bottom: 23px;
-    display: flex;
-    align-items: center;
+  margin-right: 30px;
+  margin-left: 10px;
+  margin-bottom: 23px;
+  margin-top: 10px;
+  display: flex;
+  align-items: center;
+  width: 100%;
 `;
 
 const Title = styled.div`
-    font-family: 'Pretendard';
-    font-weight:bold;
-    font-size: 1.6rem;
-`
-
-const ChallengeNumber = styled.div`
-    margin-left: auto;
-    margin-right: 2vw;
-    font-family: 'Pretendard';
-    font-size: 1.1rem;
-`
-const SeparateLine = styled.hr`
-    border: none;
-    border-top: 3px solid #000;
-    margin-right: 3vw;
-    margin-left:0px;
-    margin-top:-10px;
+  font-weight:bold;
+  font-size: 1.8rem;
+  position: fixed;
+  padding-bottom: 10px;
+  padding-top: 10px;
+  width: 100%;
+  background-color: white;
 `;
 
+const ChallengeNumber = styled.div`
+  margin-left: 240px;
+  margin-top: 10px;
+  font-size: 1.1rem;
+  position: fixed;
+`;
+
+const Line = styled.div`
+  margin-left: 10px;
+  width: 300px;
+  border-bottom: 3px solid #000;
+  position: fixed;
+`
+
 const Body = styled.div`
-    margin-top: 4vw;
-    margin-left: 0.5vw;
+  padding-top: 40px;
+  margin-left: 10px;
+  margin-right: 3.3vw; 
+  overflow-y: auto;
 `;
 
 const TestCategoryPage = () => {
@@ -142,10 +157,11 @@ const TestCategoryPage = () => {
            <TitleBox>
                 <Title>자격증/시험 챌린지</Title>
                 <ChallengeNumber>
-                    총 {total}개의 챌린지
+                    (총 {total}개)
                 </ChallengeNumber>
+                <SearchBar/>
             </TitleBox>
-            <SeparateLine/>
+            <Line/>
             <Body>
                 <RecruitList>
                     {recruit && recruit.map(challenge=>(
@@ -166,17 +182,6 @@ const TestCategoryPage = () => {
                     ))}
                 </RecruitList>
             </Body>
-        <BtnContainer>
-        <Button
-          fontSize="2.3rem"
-          title={`챌린지 생성\n \u00A0바로가기`}
-          marginTop="25px"
-          marginLeft="15px"
-          onClick={() => {
-            navigate("/agree");
-          }}
-        />
-      </BtnContainer>
         </RecruitBox>
     );
 };
